Fixed and added create todo

+++ src/main/groovy/com/thesamoanppprogrammer/GroovySpringBootTemplates/repository/TodoRepository.groovy
@@ -4,6 +4,6 @@ import com.thesamoanppprogrammer.GroovySpringBootTemplates.entity.Todo
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor
 import org.springframework.data.repository.PagingAndSortingRepository
 
-public interface TodoRepository extends PagingAndSortingRepository<Todo, Long>,
+public interface TodoRepository extends PagingAndSortingRepository<Todo, Integer>,
         JpaSpecificationExecutor<Todo> {
 }
